refactor: ExpressionParser Util 클래스로 변경, 메서드 static.
- 상태를 가지고 있을 필요가 없어 변경.
