+++ src/main/java/stringcalculator/StringCalculatorApplication.java
@@ -3,7 +3,7 @@ package stringcalculator;
 import java.util.List;
 import stringcalculator.domain.Accumulator;
 import stringcalculator.domain.Delimiters;
-import stringcalculator.domain.ExpressionParser;
+import stringcalculator.utils.ExpressionParser;
 import stringcalculator.view.InputView;
 import stringcalculator.view.ResultView;
 
@@ -13,8 +13,7 @@ public class StringCalculatorApplication {
         String delimiter = InputView.inputDelimiter();
         String expression = InputView.inputExpression();
 
-        ExpressionParser expressionParser = new ExpressionParser(new Delimiters(delimiter));
-        List<Integer> numbers = expressionParser.parse(expression);
+        List<Integer> numbers = ExpressionParser.parse(expression, new Delimiters(delimiter).getValuesToString());
 
         int result = Accumulator.calculate(numbers);
         ResultView.print(result);
